[lldb] split NativeFile in platform specific implementations (#196293)
