+++ lldb/include/module.modulemap
@@ -18,6 +18,7 @@ module lldb_Host {
   module Debug { header "lldb/Host/Debug.h" export * }
   module Editline { header "lldb/Host/Editline.h" export * }
   module FileCache { header "lldb/Host/FileCache.h" export * }
+  module FileBase { header "lldb/Host/FileBase.h" export * }
   module File { header "lldb/Host/File.h" export * }
   module FileAction { header "lldb/Host/FileAction.h" export * }
   module FileSystem { header "lldb/Host/FileSystem.h" export * }
